Move testutils to commands/utils
