+++ .github/workflows/test.yml
@@ -20,9 +20,9 @@ jobs:
         with:
           ref: ${{ github.event.pull_request.head.sha }}
       - name: Setup Go
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v3
         with:
-          go-version: "^1.17"
+          go-version: "~1.17"
       - name: Install npm
         uses: actions/setup-node@v2
         with:
@@ -44,7 +44,7 @@ jobs:
       - name: Lint
         run: go vet ./...
       - name: Staticcheck
-        uses: dominikh/staticcheck-action@v1.1.0
+        uses: dominikh/staticcheck-action@v1
 
       # Run tests
       - name: Tests
