+++ commands/utils/params_test.go
@@ -5,13 +5,12 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/jfrog/frogbot/commands/testdata"
 	"github.com/jfrog/froggit-go/vcsutils"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestExtractParamsFromEnvError(t *testing.T) {
-	unsetEnv := testdata.SetEnvAndAssert(t, map[string]string{
+	unsetEnv := SetEnvAndAssert(t, map[string]string{
 		jfrogUrlEnv:      "",
 		jfrogUserEnv:     "",
 		jfrogPasswordEnv: "",
@@ -21,13 +20,13 @@ func TestExtractParamsFromEnvError(t *testing.T) {
 	_, _, err := GetParamsAndClient()
 	assert.EqualError(t, err, "JF_URL or JF_XRAY_URL and JF_ARTIFACTORY_URL environment variables are missing")
 
-	testdata.SetEnvAndAssert(t, map[string]string{jfrogUrlEnv: "http://127.0.0.1:8081"})
+	SetEnvAndAssert(t, map[string]string{jfrogUrlEnv: "http://127.0.0.1:8081"})
 	_, _, err = GetParamsAndClient()
 	assert.EqualError(t, err, "JF_USER and JF_PASSWORD or JF_ACCESS_TOKEN environment variables are missing")
 }
 
 func TestExtractParamsFromEnvPlatform(t *testing.T) {
-	unsetEnv := testdata.SetEnvAndAssert(t, map[string]string{
+	unsetEnv := SetEnvAndAssert(t, map[string]string{
 		jfrogUrlEnv:         "http://127.0.0.1:8081",
 		jfrogUserEnv:        "admin",
 		jfrogPasswordEnv:    "password",
@@ -46,7 +45,7 @@ func TestExtractParamsFromEnvPlatform(t *testing.T) {
 }
 
 func TestExtractParamsFromEnvArtifactoryXray(t *testing.T) {
-	unsetEnv := testdata.SetEnvAndAssert(t, map[string]string{
+	unsetEnv := SetEnvAndAssert(t, map[string]string{
 		jfrogUrlEnv:            "",
 		jfrogArtifactoryUrlEnv: "http://127.0.0.1:8081/artifactory",
 		jfrogXrayUrlEnv:        "http://127.0.0.1:8081/xray",
@@ -66,7 +65,7 @@ func TestExtractParamsFromEnvArtifactoryXray(t *testing.T) {
 }
 
 func TestExtractParamsFromEnvToken(t *testing.T) {
-	unsetEnv := testdata.SetEnvAndAssert(t, map[string]string{
+	unsetEnv := SetEnvAndAssert(t, map[string]string{
 		jfrogUrlEnv:         "http://127.0.0.1:8081",
 		jfrogUserEnv:        "",
 		jfrogPasswordEnv:    "",
@@ -93,12 +92,12 @@ func TestExtractVcsProviderFromEnv(t *testing.T) {
 		assert.NoError(t, os.Unsetenv(GitProvider))
 	}()
 
-	_ = testdata.SetEnvAndAssert(t, map[string]string{GitProvider: string(GitHub)})
+	_ = SetEnvAndAssert(t, map[string]string{GitProvider: string(GitHub)})
 	vcsProvider, err := extractVcsProviderFromEnv()
 	assert.NoError(t, err)
 	assert.Equal(t, vcsutils.GitHub, vcsProvider)
 
-	_ = testdata.SetEnvAndAssert(t, map[string]string{GitProvider: string(GitLab)})
+	_ = SetEnvAndAssert(t, map[string]string{GitProvider: string(GitLab)})
 	vcsProvider, err = extractVcsProviderFromEnv()
 	assert.NoError(t, err)
 	assert.Equal(t, vcsutils.GitLab, vcsProvider)
@@ -114,19 +113,19 @@ func TestExtractInstallationCommandFromEnv(t *testing.T) {
 	assert.Empty(t, params.InstallCommandName)
 	assert.Empty(t, params.InstallCommandArgs)
 
-	_ = testdata.SetEnvAndAssert(t, map[string]string{InstallCommandEnv: "a"})
+	_ = SetEnvAndAssert(t, map[string]string{InstallCommandEnv: "a"})
 	params = &FrogbotParams{}
 	extractInstallationCommandFromEnv(params)
 	assert.Equal(t, "a", params.InstallCommandName)
 	assert.Empty(t, params.InstallCommandArgs)
 
-	_ = testdata.SetEnvAndAssert(t, map[string]string{InstallCommandEnv: "a b"})
+	_ = SetEnvAndAssert(t, map[string]string{InstallCommandEnv: "a b"})
 	params = &FrogbotParams{}
 	extractInstallationCommandFromEnv(params)
 	assert.Equal(t, "a", params.InstallCommandName)
 	assert.Equal(t, []string{"b"}, params.InstallCommandArgs)
 
-	_ = testdata.SetEnvAndAssert(t, map[string]string{InstallCommandEnv: "a b --flagName=flagValue"})
+	_ = SetEnvAndAssert(t, map[string]string{InstallCommandEnv: "a b --flagName=flagValue"})
 	params = &FrogbotParams{}
 	extractInstallationCommandFromEnv(params)
 	assert.Equal(t, "a", params.InstallCommandName)
@@ -139,32 +138,32 @@ func TestExtractGitParamsFromEnvErrors(t *testing.T) {
 	err := extractGitParamsFromEnv(params)
 	assert.EqualError(t, err, "JF_GIT_PROVIDER should be one of: 'github' or 'gitlab'")
 
-	unsetEnv1 := testdata.SetEnvAndAssert(t, map[string]string{GitProvider: "github"})
+	unsetEnv1 := SetEnvAndAssert(t, map[string]string{GitProvider: "github"})
 	defer unsetEnv1()
 	err = extractGitParamsFromEnv(params)
 	assert.EqualError(t, err, "'JF_GIT_OWNER' environment variable is missing")
 
-	unsetEnv2 := testdata.SetEnvAndAssert(t, map[string]string{GitRepoOwnerEnv: "jfrog"})
+	unsetEnv2 := SetEnvAndAssert(t, map[string]string{GitRepoOwnerEnv: "jfrog"})
 	defer unsetEnv2()
 	err = extractGitParamsFromEnv(params)
 	assert.EqualError(t, err, "'JF_GIT_REPO' environment variable is missing")
 
-	unsetEnv3 := testdata.SetEnvAndAssert(t, map[string]string{GitRepoEnv: "frogit"})
+	unsetEnv3 := SetEnvAndAssert(t, map[string]string{GitRepoEnv: "frogit"})
 	defer unsetEnv3()
 	err = extractGitParamsFromEnv(params)
 	assert.EqualError(t, err, "'JF_GIT_TOKEN' environment variable is missing")
 
-	unsetEnv4 := testdata.SetEnvAndAssert(t, map[string]string{GitTokenEnv: "123456"})
+	unsetEnv4 := SetEnvAndAssert(t, map[string]string{GitTokenEnv: "123456"})
 	defer unsetEnv4()
 	err = extractGitParamsFromEnv(params)
 	assert.EqualError(t, err, "'JF_GIT_BASE_BRANCH' environment variable is missing")
 
-	unsetEnv5 := testdata.SetEnvAndAssert(t, map[string]string{GitBaseBranchEnv: "master"})
+	unsetEnv5 := SetEnvAndAssert(t, map[string]string{GitBaseBranchEnv: "master"})
 	defer unsetEnv5()
 	err = extractGitParamsFromEnv(params)
 	assert.EqualError(t, err, "'JF_GIT_PULL_REQUEST_ID' environment variable is missing")
 
-	unsetEnv6 := testdata.SetEnvAndAssert(t, map[string]string{GitPullRequestIDEnv: "illegal-id"})
+	unsetEnv6 := SetEnvAndAssert(t, map[string]string{GitPullRequestIDEnv: "illegal-id"})
 	defer unsetEnv6()
 	err = extractGitParamsFromEnv(params)
 	_, ok := err.(*strconv.NumError)
